requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
PyJWT==2.9.0
cryptography==43.0.1
python-dotenv==1.0.1
pydantic[email]==2.9.2
supabase==2.9.1
openai==1.51.0
mistune==3.0.2
deepgram-sdk==3.7.7
elevenlabs==1.9.0
numpy==1.26.4
websockets==13.1
python-multipart==0.0.12
